List Calendar endpoints

Returns a list of selected calendar API v3 endpoints, as stored inside the googlecalendar package. The names of this list (or the id sub-elements) are the nicknames that can be used to specify an endpoint in request_generate(). For each endpoint, we store its nickname or id, the associated HTTP verb, the path, and details about the parameters. This list is derived programmatically from the calendar API v3 Discovery Document.

calendar_endpoints(i = NULL)

Arguments

i
The name(s) or integer index(ices) of the endpoints to return. Optional. By default, the entire list is returned.

Value

A list containing some or all of the subset of the calendar API v3 endpoints that are used internally by gcalendr.

Examples

str(calendar_endpoints(), max.level = 2)

#> List of 37
#>  $ calendar.acl.delete         :List of 7
#>   ..$ id            : chr "calendar.acl.delete"
#>   ..$ httpMethod    : chr "DELETE"
#>   ..$ path          : 'fs_path' chr "calendar/v3/calendars/{calendarId}/acl/{ruleId}"
#>   ..$ parameters    :List of 9
#>   ..$ scopes        : chr "calendar"
#>   ..$ description   : chr "Deletes an access control rule."
#>   ..$ parameterOrder: chr [1:2] "calendarId" "ruleId"
#>  $ calendar.acl.get            :List of 8
#>   ..$ id            : chr "calendar.acl.get"
#>   ..$ httpMethod    : chr "GET"
#>   ..$ path          : 'fs_path' chr "calendar/v3/calendars/{calendarId}/acl/{ruleId}"
#>   ..$ parameters    :List of 9
#>   ..$ scopes        : chr "calendar, calendar.readonly"
#>   ..$ description   : chr "Returns an access control rule."
#>   ..$ response      : chr "AclRule"
#>   ..$ parameterOrder: chr [1:2] "calendarId" "ruleId"
#>  $ calendar.acl.insert         :List of 9
#>   ..$ id            : chr "calendar.acl.insert"
#>   ..$ httpMethod    : chr "POST"
#>   ..$ path          : 'fs_path' chr "calendar/v3/calendars/{calendarId}/acl"
#>   ..$ parameters    :List of 14
#>   ..$ scopes        : chr "calendar"
#>   ..$ description   : chr "Creates an access control rule."
#>   ..$ request       : chr "AclRule"
#>   ..$ response      : chr "AclRule"
#>   ..$ parameterOrder: chr "calendarId"
#>  $ calendar.acl.list           :List of 9
#>   ..$ id                  : chr "calendar.acl.list"
#>   ..$ httpMethod          : chr "GET"
#>   ..$ path                : 'fs_path' chr "calendar/v3/calendars/{calendarId}/acl"
#>   ..$ parameters          :List of 12
#>   ..$ scopes              : chr "calendar"
#>   ..$ description         : chr "Returns the rules in the access control list for the calendar."
#>   ..$ response            : chr "Acl"
#>   ..$ parameterOrder      : chr "calendarId"
#>   ..$ supportsSubscription: logi TRUE
#>  $ calendar.acl.patch          :List of 9
#>   ..$ id            : chr "calendar.acl.patch"
#>   ..$ httpMethod    : chr "PATCH"
#>   ..$ path          : 'fs_path' chr "calendar/v3/calendars/{calendarId}/acl/{ruleId}"
#>   ..$ parameters    :List of 15
#>   ..$ scopes        : chr "calendar"
#>   ..$ description   : chr "Updates an access control rule. This method supports patch semantics."
#>   ..$ request       : chr "AclRule"
#>   ..$ response      : chr "AclRule"
#>   ..$ parameterOrder: chr [1:2] "calendarId" "ruleId"
#>  $ calendar.acl.update         :List of 9
#>   ..$ id            : chr "calendar.acl.update"
#>   ..$ httpMethod    : chr "PUT"
#>   ..$ path          : 'fs_path' chr "calendar/v3/calendars/{calendarId}/acl/{ruleId}"
#>   ..$ parameters    :List of 15
#>   ..$ scopes        : chr "calendar"
#>   ..$ description   : chr "Updates an access control rule."
#>   ..$ request       : chr "AclRule"
#>   ..$ response      : chr "AclRule"
#>   ..$ parameterOrder: chr [1:2] "calendarId" "ruleId"
#>  $ calendar.acl.watch          :List of 10
#>   ..$ id                  : chr "calendar.acl.watch"
#>   ..$ httpMethod          : chr "POST"
#>   ..$ path                : 'fs_path' chr "calendar/v3/calendars/{calendarId}/acl/watch"
#>   ..$ parameters          :List of 22
#>   ..$ scopes              : chr "calendar"
#>   ..$ description         : chr "Watch for changes to ACL resources."
#>   ..$ request             : chr "Channel"
#>   ..$ response            : chr "Channel"
#>   ..$ parameterOrder      : chr "calendarId"
#>   ..$ supportsSubscription: logi TRUE
#>  $ calendar.calendarList.delete:List of 7
#>   ..$ id            : chr "calendar.calendarList.delete"
#>   ..$ httpMethod    : chr "DELETE"
#>   ..$ path          : 'fs_path' chr "calendar/v3/users/me/calendarList/{calendarId}"
#>   ..$ parameters    :List of 8
#>   ..$ scopes        : chr "calendar"
#>   ..$ description   : chr "Removes a calendar from the user's calendar list."
#>   ..$ parameterOrder: chr "calendarId"
#>  $ calendar.calendarList.get   :List of 8
#>   ..$ id            : chr "calendar.calendarList.get"
#>   ..$ httpMethod    : chr "GET"
#>   ..$ path          : 'fs_path' chr "calendar/v3/users/me/calendarList/{calendarId}"
#>   ..$ parameters    :List of 8
#>   ..$ scopes        : chr "calendar, calendar.readonly"
#>   ..$ description   : chr "Returns a calendar from the user's calendar list."
#>   ..$ response      : chr "CalendarListEntry"
#>   ..$ parameterOrder: chr "calendarId"
#>  $ calendar.calendarList.insert:List of 8
#>   ..$ id         : chr "calendar.calendarList.insert"
#>   ..$ httpMethod : chr "POST"
#>   ..$ path       : 'fs_path' chr "calendar/v3/users/me/calendarList"
#>   ..$ parameters :List of 27
#>   ..$ scopes     : chr "calendar"
#>   ..$ description: chr "Inserts an existing calendar into the user's calendar list."
#>   ..$ request    : chr "CalendarListEntry"
#>   ..$ response   : chr "CalendarListEntry"
#>  $ calendar.calendarList.list  :List of 8
#>   ..$ id                  : chr "calendar.calendarList.list"
#>   ..$ httpMethod          : chr "GET"
#>   ..$ path                : 'fs_path' chr "calendar/v3/users/me/calendarList"
#>   ..$ parameters          :List of 13
#>   ..$ scopes              : chr "calendar, calendar.readonly"
#>   ..$ description         : chr "Returns the calendars on the user's calendar list."
#>   ..$ response            : chr "CalendarList"
#>   ..$ supportsSubscription: logi TRUE
#>  $ calendar.calendarList.patch :List of 9
#>   ..$ id            : chr "calendar.calendarList.patch"
#>   ..$ httpMethod    : chr "PATCH"
#>   ..$ path          : 'fs_path' chr "calendar/v3/users/me/calendarList/{calendarId}"
#>   ..$ parameters    :List of 28
#>   ..$ scopes        : chr "calendar"
#>   ..$ description   : chr "Updates an existing calendar on the user's calendar list. This method supports patch semantics."
#>   ..$ request       : chr "CalendarListEntry"
#>   ..$ response      : chr "CalendarListEntry"
#>   ..$ parameterOrder: chr "calendarId"
#>  $ calendar.calendarList.update:List of 9
#>   ..$ id            : chr "calendar.calendarList.update"
#>   ..$ httpMethod    : chr "PUT"
#>   ..$ path          : 'fs_path' chr "calendar/v3/users/me/calendarList/{calendarId}"
#>   ..$ parameters    :List of 28
#>   ..$ scopes        : chr "calendar"
#>   ..$ description   : chr "Updates an existing calendar on the user's calendar list."
#>   ..$ request       : chr "CalendarListEntry"
#>   ..$ response      : chr "CalendarListEntry"
#>   ..$ parameterOrder: chr "calendarId"
#>  $ calendar.calendarList.watch :List of 9
#>   ..$ id                  : chr "calendar.calendarList.watch"
#>   ..$ httpMethod          : chr "POST"
#>   ..$ path                : 'fs_path' chr "calendar/v3/users/me/calendarList/watch"
#>   ..$ parameters          :List of 23
#>   ..$ scopes              : chr "calendar, calendar.readonly"
#>   ..$ description         : chr "Watch for changes to CalendarList resources."
#>   ..$ request             : chr "Channel"
#>   ..$ response            : chr "Channel"
#>   ..$ supportsSubscription: logi TRUE
#>  $ calendar.calendars.clear    :List of 7
#>   ..$ id            : chr "calendar.calendars.clear"
#>   ..$ httpMethod    : chr "POST"
#>   ..$ path          : 'fs_path' chr "calendar/v3/calendars/{calendarId}/clear"
#>   ..$ parameters    :List of 8
#>   ..$ scopes        : chr "calendar"
#>   ..$ description   : chr "Clears a primary calendar. This operation deletes all events associated with the primary calendar of an account."
#>   ..$ parameterOrder: chr "calendarId"
#>  $ calendar.calendars.delete   :List of 7
#>   ..$ id            : chr "calendar.calendars.delete"
#>   ..$ httpMethod    : chr "DELETE"
#>   ..$ path          : 'fs_path' chr "calendar/v3/calendars/{calendarId}"
#>   ..$ parameters    :List of 8
#>   ..$ scopes        : chr "calendar"
#>   ..$ description   : chr "Deletes a secondary calendar. Use calendars.clear for clearing all events on primary calendars."
#>   ..$ parameterOrder: chr "calendarId"
#>  $ calendar.calendars.get      :List of 8
#>   ..$ id            : chr "calendar.calendars.get"
#>   ..$ httpMethod    : chr "GET"
#>   ..$ path          : 'fs_path' chr "calendar/v3/calendars/{calendarId}"
#>   ..$ parameters    :List of 8
#>   ..$ scopes        : chr "calendar, calendar.readonly"
#>   ..$ description   : chr "Returns metadata for a calendar."
#>   ..$ response      : chr "Calendar"
#>   ..$ parameterOrder: chr "calendarId"
#>  $ calendar.calendars.insert   :List of 8
#>   ..$ id         : chr "calendar.calendars.insert"
#>   ..$ httpMethod : chr "POST"
#>   ..$ path       : 'fs_path' chr "calendar/v3/calendars"
#>   ..$ scopes     : chr "calendar"
#>   ..$ description: chr "Creates a secondary calendar."
#>   ..$ request    : chr "Calendar"
#>   ..$ response   : chr "Calendar"
#>   ..$ parameters :List of 15
#>  $ calendar.calendars.patch    :List of 9
#>   ..$ id            : chr "calendar.calendars.patch"
#>   ..$ httpMethod    : chr "PATCH"
#>   ..$ path          : 'fs_path' chr "calendar/v3/calendars/{calendarId}"
#>   ..$ parameters    :List of 16
#>   ..$ scopes        : chr "calendar"
#>   ..$ description   : chr "Updates metadata for a calendar. This method supports patch semantics."
#>   ..$ request       : chr "Calendar"
#>   ..$ response      : chr "Calendar"
#>   ..$ parameterOrder: chr "calendarId"
#>  $ calendar.calendars.update   :List of 9
#>   ..$ id            : chr "calendar.calendars.update"
#>   ..$ httpMethod    : chr "PUT"
#>   ..$ path          : 'fs_path' chr "calendar/v3/calendars/{calendarId}"
#>   ..$ parameters    :List of 16
#>   ..$ scopes        : chr "calendar"
#>   ..$ description   : chr "Updates metadata for a calendar."
#>   ..$ request       : chr "Calendar"
#>   ..$ response      : chr "Calendar"
#>   ..$ parameterOrder: chr "calendarId"
#>  $ calendar.channels.stop      :List of 7
#>   ..$ id         : chr "calendar.channels.stop"
#>   ..$ httpMethod : chr "POST"
#>   ..$ path       : 'fs_path' chr "calendar/v3/channels/stop"
#>   ..$ scopes     : chr "calendar, calendar.events, calendar.events.readonly, calendar.readonly, calendar.settings.readonly"
#>   ..$ description: chr "Stop watching resources through this channel"
#>   ..$ request    : chr "Channel"
#>   ..$ parameters :List of 17
#>  $ calendar.colors.get         :List of 7
#>   ..$ id         : chr "calendar.colors.get"
#>   ..$ httpMethod : chr "GET"
#>   ..$ path       : 'fs_path' chr "calendar/v3/colors"
#>   ..$ scopes     : chr "calendar, calendar.readonly"
#>   ..$ description: chr "Returns the color definitions for calendars and events."
#>   ..$ response   : chr "Colors"
#>   ..$ parameters :List of 7
#>  $ calendar.events.delete      :List of 7
#>   ..$ id            : chr "calendar.events.delete"
#>   ..$ httpMethod    : chr "DELETE"
#>   ..$ path          : 'fs_path' chr "calendar/v3/calendars/{calendarId}/events/{eventId}"
#>   ..$ parameters    :List of 11
#>   ..$ scopes        : chr "calendar, calendar.events"
#>   ..$ description   : chr "Deletes an event."
#>   ..$ parameterOrder: chr [1:2] "calendarId" "eventId"
#>  $ calendar.events.get         :List of 8
#>   ..$ id            : chr "calendar.events.get"
#>   ..$ httpMethod    : chr "GET"
#>   ..$ path          : 'fs_path' chr "calendar/v3/calendars/{calendarId}/events/{eventId}"
#>   ..$ parameters    :List of 12
#>   ..$ scopes        : chr "calendar, calendar.events, calendar.events.readonly, calendar.readonly"
#>   ..$ description   : chr "Returns an event."
#>   ..$ response      : chr "Event"
#>   ..$ parameterOrder: chr [1:2] "calendarId" "eventId"
#>  $ calendar.events.import      :List of 9
#>   ..$ id            : chr "calendar.events.import"
#>   ..$ httpMethod    : chr "POST"
#>   ..$ path          : 'fs_path' chr "calendar/v3/calendars/{calendarId}/events/import"
#>   ..$ parameters    :List of 48
#>   ..$ scopes        : chr "calendar, calendar.events"
#>   ..$ description   : chr "Imports an event. This operation is used to add a private copy of an existing event to a calendar."
#>   ..$ request       : chr "Event"
#>   ..$ response      : chr "Event"
#>   ..$ parameterOrder: chr "calendarId"
#>  $ calendar.events.insert      :List of 9
#>   ..$ id            : chr "calendar.events.insert"
#>   ..$ httpMethod    : chr "POST"
#>   ..$ path          : 'fs_path' chr "calendar/v3/calendars/{calendarId}/events"
#>   ..$ parameters    :List of 51
#>   ..$ scopes        : chr "calendar, calendar.events"
#>   ..$ description   : chr "Creates an event."
#>   ..$ request       : chr "Event"
#>   ..$ response      : chr "Event"
#>   ..$ parameterOrder: chr "calendarId"
#>  $ calendar.events.instances   :List of 9
#>   ..$ id                  : chr "calendar.events.instances"
#>   ..$ httpMethod          : chr "GET"
#>   ..$ path                : 'fs_path' chr "calendar/v3/calendars/{calendarId}/events/{eventId}/instances"
#>   ..$ parameters          :List of 18
#>   ..$ scopes              : chr "calendar, calendar.events, calendar.events.readonly, calendar.readonly"
#>   ..$ description         : chr "Returns instances of the specified recurring event."
#>   ..$ response            : chr "Events"
#>   ..$ parameterOrder      : chr [1:2] "calendarId" "eventId"
#>   ..$ supportsSubscription: logi TRUE
#>  $ calendar.events.list        :List of 9
#>   ..$ id                  : chr "calendar.events.list"
#>   ..$ httpMethod          : chr "GET"
#>   ..$ path                : 'fs_path' chr "calendar/v3/calendars/{calendarId}/events"
#>   ..$ parameters          :List of 25
#>   ..$ scopes              : chr "calendar, calendar.events, calendar.events.readonly, calendar.readonly"
#>   ..$ description         : chr "Returns events on the specified calendar."
#>   ..$ response            : chr "Events"
#>   ..$ parameterOrder      : chr "calendarId"
#>   ..$ supportsSubscription: logi TRUE
#>  $ calendar.events.move        :List of 8
#>   ..$ id            : chr "calendar.events.move"
#>   ..$ httpMethod    : chr "POST"
#>   ..$ path          : 'fs_path' chr "calendar/v3/calendars/{calendarId}/events/{eventId}/move"
#>   ..$ parameters    :List of 12
#>   ..$ scopes        : chr "calendar, calendar.events"
#>   ..$ description   : chr "Moves an event to another calendar, i.e. changes an event's organizer."
#>   ..$ response      : chr "Event"
#>   ..$ parameterOrder: chr [1:3] "calendarId" "eventId" "destination"
#>  $ calendar.events.patch       :List of 9
#>   ..$ id            : chr "calendar.events.patch"
#>   ..$ httpMethod    : chr "PATCH"
#>   ..$ path          : 'fs_path' chr "calendar/v3/calendars/{calendarId}/events/{eventId}"
#>   ..$ parameters    :List of 53
#>   ..$ scopes        : chr "calendar, calendar.events"
#>   ..$ description   : chr "Updates an event. This method supports patch semantics."
#>   ..$ request       : chr "Event"
#>   ..$ response      : chr "Event"
#>   ..$ parameterOrder: chr [1:2] "calendarId" "eventId"
#>  $ calendar.events.quickAdd    :List of 8
#>   ..$ id            : chr "calendar.events.quickAdd"
#>   ..$ httpMethod    : chr "POST"
#>   ..$ path          : 'fs_path' chr "calendar/v3/calendars/{calendarId}/events/quickAdd"
#>   ..$ parameters    :List of 11
#>   ..$ scopes        : chr "calendar, calendar.events"
#>   ..$ description   : chr "Creates an event based on a simple text string."
#>   ..$ response      : chr "Event"
#>   ..$ parameterOrder: chr [1:2] "calendarId" "text"
#>  $ calendar.events.update      :List of 9
#>   ..$ id            : chr "calendar.events.update"
#>   ..$ httpMethod    : chr "PUT"
#>   ..$ path          : 'fs_path' chr "calendar/v3/calendars/{calendarId}/events/{eventId}"
#>   ..$ parameters    :List of 53
#>   ..$ scopes        : chr "calendar, calendar.events"
#>   ..$ description   : chr "Updates an event."
#>   ..$ request       : chr "Event"
#>   ..$ response      : chr "Event"
#>   ..$ parameterOrder: chr [1:2] "calendarId" "eventId"
#>  $ calendar.events.watch       :List of 10
#>   ..$ id                  : chr "calendar.events.watch"
#>   ..$ httpMethod          : chr "POST"
#>   ..$ path                : 'fs_path' chr "calendar/v3/calendars/{calendarId}/events/watch"
#>   ..$ parameters          :List of 35
#>   ..$ scopes              : chr "calendar, calendar.events, calendar.events.readonly, calendar.readonly"
#>   ..$ description         : chr "Watch for changes to Events resources."
#>   ..$ request             : chr "Channel"
#>   ..$ response            : chr "Channel"
#>   ..$ parameterOrder      : chr "calendarId"
#>   ..$ supportsSubscription: logi TRUE
#>  $ calendar.freebusy.query     :List of 8
#>   ..$ id         : chr "calendar.freebusy.query"
#>   ..$ httpMethod : chr "POST"
#>   ..$ path       : 'fs_path' chr "calendar/v3/freeBusy"
#>   ..$ scopes     : chr "calendar, calendar.readonly"
#>   ..$ description: chr "Returns free/busy information for a set of calendars."
#>   ..$ request    : chr "FreeBusyRequest"
#>   ..$ response   : chr "FreeBusyResponse"
#>   ..$ parameters :List of 13
#>  $ calendar.settings.get       :List of 8
#>   ..$ id            : chr "calendar.settings.get"
#>   ..$ httpMethod    : chr "GET"
#>   ..$ path          : 'fs_path' chr "calendar/v3/users/me/settings/{setting}"
#>   ..$ parameters    :List of 8
#>   ..$ scopes        : chr "calendar, calendar.readonly, calendar.settings.readonly"
#>   ..$ description   : chr "Returns a single user setting."
#>   ..$ response      : chr "Setting"
#>   ..$ parameterOrder: chr "setting"
#>  $ calendar.settings.list      :List of 8
#>   ..$ id                  : chr "calendar.settings.list"
#>   ..$ httpMethod          : chr "GET"
#>   ..$ path                : 'fs_path' chr "calendar/v3/users/me/settings"
#>   ..$ parameters          :List of 10
#>   ..$ scopes              : chr "calendar, calendar.readonly, calendar.settings.readonly"
#>   ..$ description         : chr "Returns all user settings for the authenticated user."
#>   ..$ response            : chr "Settings"
#>   ..$ supportsSubscription: logi TRUE
#>  $ calendar.settings.watch     :List of 9
#>   ..$ id                  : chr "calendar.settings.watch"
#>   ..$ httpMethod          : chr "POST"
#>   ..$ path                : 'fs_path' chr "calendar/v3/users/me/settings/watch"
#>   ..$ parameters          :List of 20
#>   ..$ scopes              : chr "calendar, calendar.readonly, calendar.settings.readonly"
#>   ..$ description         : chr "Watch for changes to Settings resources."
#>   ..$ request             : chr "Channel"
#>   ..$ response            : chr "Channel"
#>   ..$ supportsSubscription: logi TRUE

calendar_endpoints("calendar.events.list")

#> $calendar.events.list
#> $calendar.events.list$id
#> [1] "calendar.events.list"
#> 
#> $calendar.events.list$httpMethod
#> [1] "GET"
#> 
#> $calendar.events.list$path
#> calendar/v3/calendars/{calendarId}/events
#> 
#> $calendar.events.list$parameters
#> $calendar.events.list$parameters$alwaysIncludeEmail
#> $calendar.events.list$parameters$alwaysIncludeEmail$type
#> [1] "boolean"
#> 
#> $calendar.events.list$parameters$alwaysIncludeEmail$description
#> [1] "Whether to always include a value in the email field for the organizer, creator and attendees, even if no real email is available (i.e. a generated, non-working value will be provided). The use of this option is discouraged and should only be used by clients which cannot handle the absence of an email address value in the mentioned places. Optional. The default is False."
#> 
#> $calendar.events.list$parameters$alwaysIncludeEmail$location
#> [1] "query"
#> 
#> 
#> $calendar.events.list$parameters$calendarId
#> $calendar.events.list$parameters$calendarId$type
#> [1] "string"
#> 
#> $calendar.events.list$parameters$calendarId$description
#> [1] "Calendar identifier. To retrieve calendar IDs call the calendarList.list method. If you want to access the primary calendar of the currently logged in user, use the \"primary\" keyword."
#> 
#> $calendar.events.list$parameters$calendarId$required
#> [1] TRUE
#> 
#> $calendar.events.list$parameters$calendarId$location
#> [1] "path"
#> 
#> 
#> $calendar.events.list$parameters$iCalUID
#> $calendar.events.list$parameters$iCalUID$type
#> [1] "string"
#> 
#> $calendar.events.list$parameters$iCalUID$description
#> [1] "Specifies event ID in the iCalendar format to be included in the response. Optional."
#> 
#> $calendar.events.list$parameters$iCalUID$location
#> [1] "query"
#> 
#> 
#> $calendar.events.list$parameters$maxAttendees
#> $calendar.events.list$parameters$maxAttendees$type
#> [1] "integer"
#> 
#> $calendar.events.list$parameters$maxAttendees$description
#> [1] "The maximum number of attendees to include in the response. If there are more than the specified number of attendees, only the participant is returned. Optional."
#> 
#> $calendar.events.list$parameters$maxAttendees$format
#> [1] "int32"
#> 
#> $calendar.events.list$parameters$maxAttendees$minimum
#> [1] "1"
#> 
#> $calendar.events.list$parameters$maxAttendees$location
#> [1] "query"
#> 
#> 
#> $calendar.events.list$parameters$maxResults
#> $calendar.events.list$parameters$maxResults$type
#> [1] "integer"
#> 
#> $calendar.events.list$parameters$maxResults$description
#> [1] "Maximum number of events returned on one result page. The number of events in the resulting page may be less than this value, or none at all, even if there are more events matching the query. Incomplete pages can be detected by a non-empty nextPageToken field in the response. By default the value is 250 events. The page size can never be larger than 2500 events. Optional."
#> 
#> $calendar.events.list$parameters$maxResults$default
#> [1] "250"
#> 
#> $calendar.events.list$parameters$maxResults$format
#> [1] "int32"
#> 
#> $calendar.events.list$parameters$maxResults$minimum
#> [1] "1"
#> 
#> $calendar.events.list$parameters$maxResults$location
#> [1] "query"
#> 
#> 
#> $calendar.events.list$parameters$orderBy
#> $calendar.events.list$parameters$orderBy$type
#> [1] "string"
#> 
#> $calendar.events.list$parameters$orderBy$description
#> [1] "The order of the events returned in the result. Optional. The default is an unspecified, stable order."
#> 
#> $calendar.events.list$parameters$orderBy$enum
#> [1] "startTime" "updated"  
#> 
#> $calendar.events.list$parameters$orderBy$enumDescriptions
#> [1] "Order by the start date/time (ascending). This is only available when querying single events (i.e. the parameter singleEvents is True)"
#> [2] "Order by last modification time (ascending)."                                                                                          
#> 
#> $calendar.events.list$parameters$orderBy$location
#> [1] "query"
#> 
#> 
#> $calendar.events.list$parameters$pageToken
#> $calendar.events.list$parameters$pageToken$type
#> [1] "string"
#> 
#> $calendar.events.list$parameters$pageToken$description
#> [1] "Token specifying which result page to return. Optional."
#> 
#> $calendar.events.list$parameters$pageToken$location
#> [1] "query"
#> 
#> 
#> $calendar.events.list$parameters$privateExtendedProperty
#> $calendar.events.list$parameters$privateExtendedProperty$type
#> [1] "string"
#> 
#> $calendar.events.list$parameters$privateExtendedProperty$description
#> [1] "Extended properties constraint specified as propertyName=value. Matches only private properties. This parameter might be repeated multiple times to return events that match all given constraints."
#> 
#> $calendar.events.list$parameters$privateExtendedProperty$repeated
#> [1] TRUE
#> 
#> $calendar.events.list$parameters$privateExtendedProperty$location
#> [1] "query"
#> 
#> 
#> $calendar.events.list$parameters$q
#> $calendar.events.list$parameters$q$type
#> [1] "string"
#> 
#> $calendar.events.list$parameters$q$description
#> [1] "Free text search terms to find events that match these terms in any field, except for extended properties. Optional."
#> 
#> $calendar.events.list$parameters$q$location
#> [1] "query"
#> 
#> 
#> $calendar.events.list$parameters$sharedExtendedProperty
#> $calendar.events.list$parameters$sharedExtendedProperty$type
#> [1] "string"
#> 
#> $calendar.events.list$parameters$sharedExtendedProperty$description
#> [1] "Extended properties constraint specified as propertyName=value. Matches only shared properties. This parameter might be repeated multiple times to return events that match all given constraints."
#> 
#> $calendar.events.list$parameters$sharedExtendedProperty$repeated
#> [1] TRUE
#> 
#> $calendar.events.list$parameters$sharedExtendedProperty$location
#> [1] "query"
#> 
#> 
#> $calendar.events.list$parameters$showDeleted
#> $calendar.events.list$parameters$showDeleted$type
#> [1] "boolean"
#> 
#> $calendar.events.list$parameters$showDeleted$description
#> [1] "Whether to include deleted events (with status equals \"cancelled\") in the result. Cancelled instances of recurring events (but not the underlying recurring event) will still be included if showDeleted and singleEvents are both False. If showDeleted and singleEvents are both True, only single instances of deleted events (but not the underlying recurring events) are returned. Optional. The default is False."
#> 
#> $calendar.events.list$parameters$showDeleted$location
#> [1] "query"
#> 
#> 
#> $calendar.events.list$parameters$showHiddenInvitations
#> $calendar.events.list$parameters$showHiddenInvitations$type
#> [1] "boolean"
#> 
#> $calendar.events.list$parameters$showHiddenInvitations$description
#> [1] "Whether to include hidden invitations in the result. Optional. The default is False."
#> 
#> $calendar.events.list$parameters$showHiddenInvitations$location
#> [1] "query"
#> 
#> 
#> $calendar.events.list$parameters$singleEvents
#> $calendar.events.list$parameters$singleEvents$type
#> [1] "boolean"
#> 
#> $calendar.events.list$parameters$singleEvents$description
#> [1] "Whether to expand recurring events into instances and only return single one-off events and instances of recurring events, but not the underlying recurring events themselves. Optional. The default is False."
#> 
#> $calendar.events.list$parameters$singleEvents$location
#> [1] "query"
#> 
#> 
#> $calendar.events.list$parameters$syncToken
#> $calendar.events.list$parameters$syncToken$type
#> [1] "string"
#> 
#> $calendar.events.list$parameters$syncToken$description
#> [1] "Token obtained from the nextSyncToken field returned on the last page of results from the previous list request. It makes the result of this list request contain only entries that have changed since then. All events deleted since the previous list request will always be in the result set and it is not allowed to set showDeleted to False.\nThere are several query parameters that cannot be specified together with nextSyncToken to ensure consistency of the client state.\n\nThese are: \n- iCalUID \n- orderBy \n- privateExtendedProperty \n- q \n- sharedExtendedProperty \n- timeMin \n- timeMax \n- updatedMin If the syncToken expires, the server will respond with a 410 GONE response code and the client should clear its storage and perform a full synchronization without any syncToken.\nLearn more about incremental synchronization.\nOptional. The default is to return all entries."
#> 
#> $calendar.events.list$parameters$syncToken$location
#> [1] "query"
#> 
#> 
#> $calendar.events.list$parameters$timeMax
#> $calendar.events.list$parameters$timeMax$type
#> [1] "string"
#> 
#> $calendar.events.list$parameters$timeMax$description
#> [1] "Upper bound (exclusive) for an event's start time to filter by. Optional. The default is not to filter by start time. Must be an RFC3339 timestamp with mandatory time zone offset, for example, 2011-06-03T10:00:00-07:00, 2011-06-03T10:00:00Z. Milliseconds may be provided but are ignored. If timeMin is set, timeMax must be greater than timeMin."
#> 
#> $calendar.events.list$parameters$timeMax$format
#> [1] "date-time"
#> 
#> $calendar.events.list$parameters$timeMax$location
#> [1] "query"
#> 
#> 
#> $calendar.events.list$parameters$timeMin
#> $calendar.events.list$parameters$timeMin$type
#> [1] "string"
#> 
#> $calendar.events.list$parameters$timeMin$description
#> [1] "Lower bound (exclusive) for an event's end time to filter by. Optional. The default is not to filter by end time. Must be an RFC3339 timestamp with mandatory time zone offset, for example, 2011-06-03T10:00:00-07:00, 2011-06-03T10:00:00Z. Milliseconds may be provided but are ignored. If timeMax is set, timeMin must be smaller than timeMax."
#> 
#> $calendar.events.list$parameters$timeMin$format
#> [1] "date-time"
#> 
#> $calendar.events.list$parameters$timeMin$location
#> [1] "query"
#> 
#> 
#> $calendar.events.list$parameters$timeZone
#> $calendar.events.list$parameters$timeZone$type
#> [1] "string"
#> 
#> $calendar.events.list$parameters$timeZone$description
#> [1] "Time zone used in the response. Optional. The default is the time zone of the calendar."
#> 
#> $calendar.events.list$parameters$timeZone$location
#> [1] "query"
#> 
#> 
#> $calendar.events.list$parameters$updatedMin
#> $calendar.events.list$parameters$updatedMin$type
#> [1] "string"
#> 
#> $calendar.events.list$parameters$updatedMin$description
#> [1] "Lower bound for an event's last modification time (as a RFC3339 timestamp) to filter by. When specified, entries deleted since this time will always be included regardless of showDeleted. Optional. The default is not to filter by last modification time."
#> 
#> $calendar.events.list$parameters$updatedMin$format
#> [1] "date-time"
#> 
#> $calendar.events.list$parameters$updatedMin$location
#> [1] "query"
#> 
#> 
#> $calendar.events.list$parameters$alt
#> $calendar.events.list$parameters$alt$type
#> [1] "string"
#> 
#> $calendar.events.list$parameters$alt$description
#> [1] "Data format for the response."
#> 
#> $calendar.events.list$parameters$alt$default
#> [1] "json"
#> 
#> $calendar.events.list$parameters$alt$enum
#> [1] "json"
#> 
#> $calendar.events.list$parameters$alt$enumDescriptions
#> [1] "Responses with Content-Type of application/json"
#> 
#> $calendar.events.list$parameters$alt$location
#> [1] "query"
#> 
#> 
#> $calendar.events.list$parameters$fields
#> $calendar.events.list$parameters$fields$type
#> [1] "string"
#> 
#> $calendar.events.list$parameters$fields$description
#> [1] "Selector specifying which fields to include in a partial response."
#> 
#> $calendar.events.list$parameters$fields$location
#> [1] "query"
#> 
#> 
#> $calendar.events.list$parameters$key
#> $calendar.events.list$parameters$key$type
#> [1] "string"
#> 
#> $calendar.events.list$parameters$key$description
#> [1] "API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token."
#> 
#> $calendar.events.list$parameters$key$location
#> [1] "query"
#> 
#> 
#> $calendar.events.list$parameters$oauth_token
#> $calendar.events.list$parameters$oauth_token$type
#> [1] "string"
#> 
#> $calendar.events.list$parameters$oauth_token$description
#> [1] "OAuth 2.0 token for the current user."
#> 
#> $calendar.events.list$parameters$oauth_token$location
#> [1] "query"
#> 
#> 
#> $calendar.events.list$parameters$prettyPrint
#> $calendar.events.list$parameters$prettyPrint$type
#> [1] "boolean"
#> 
#> $calendar.events.list$parameters$prettyPrint$description
#> [1] "Returns response with indentations and line breaks."
#> 
#> $calendar.events.list$parameters$prettyPrint$default
#> [1] "true"
#> 
#> $calendar.events.list$parameters$prettyPrint$location
#> [1] "query"
#> 
#> 
#> $calendar.events.list$parameters$quotaUser
#> $calendar.events.list$parameters$quotaUser$type
#> [1] "string"
#> 
#> $calendar.events.list$parameters$quotaUser$description
#> [1] "An opaque string that represents a user for quota purposes. Must not exceed 40 characters."
#> 
#> $calendar.events.list$parameters$quotaUser$location
#> [1] "query"
#> 
#> 
#> $calendar.events.list$parameters$userIp
#> $calendar.events.list$parameters$userIp$type
#> [1] "string"
#> 
#> $calendar.events.list$parameters$userIp$description
#> [1] "Deprecated. Please use quotaUser instead."
#> 
#> $calendar.events.list$parameters$userIp$location
#> [1] "query"
#> 
#> 
#> 
#> $calendar.events.list$scopes
#> [1] "calendar, calendar.events, calendar.events.readonly, calendar.readonly"
#> 
#> $calendar.events.list$description
#> [1] "Returns events on the specified calendar."
#> 
#> $calendar.events.list$response
#> [1] "Events"
#> 
#> $calendar.events.list$parameterOrder
#> [1] "calendarId"
#> 
#> $calendar.events.list$supportsSubscription
#> [1] TRUE
#> 
#> 

calendar_endpoints(4)

#> $calendar.acl.list
#> $calendar.acl.list$id
#> [1] "calendar.acl.list"
#> 
#> $calendar.acl.list$httpMethod
#> [1] "GET"
#> 
#> $calendar.acl.list$path
#> calendar/v3/calendars/{calendarId}/acl
#> 
#> $calendar.acl.list$parameters
#> $calendar.acl.list$parameters$calendarId
#> $calendar.acl.list$parameters$calendarId$type
#> [1] "string"
#> 
#> $calendar.acl.list$parameters$calendarId$description
#> [1] "Calendar identifier. To retrieve calendar IDs call the calendarList.list method. If you want to access the primary calendar of the currently logged in user, use the \"primary\" keyword."
#> 
#> $calendar.acl.list$parameters$calendarId$required
#> [1] TRUE
#> 
#> $calendar.acl.list$parameters$calendarId$location
#> [1] "path"
#> 
#> 
#> $calendar.acl.list$parameters$maxResults
#> $calendar.acl.list$parameters$maxResults$type
#> [1] "integer"
#> 
#> $calendar.acl.list$parameters$maxResults$description
#> [1] "Maximum number of entries returned on one result page. By default the value is 100 entries. The page size can never be larger than 250 entries. Optional."
#> 
#> $calendar.acl.list$parameters$maxResults$format
#> [1] "int32"
#> 
#> $calendar.acl.list$parameters$maxResults$minimum
#> [1] "1"
#> 
#> $calendar.acl.list$parameters$maxResults$location
#> [1] "query"
#> 
#> 
#> $calendar.acl.list$parameters$pageToken
#> $calendar.acl.list$parameters$pageToken$type
#> [1] "string"
#> 
#> $calendar.acl.list$parameters$pageToken$description
#> [1] "Token specifying which result page to return. Optional."
#> 
#> $calendar.acl.list$parameters$pageToken$location
#> [1] "query"
#> 
#> 
#> $calendar.acl.list$parameters$showDeleted
#> $calendar.acl.list$parameters$showDeleted$type
#> [1] "boolean"
#> 
#> $calendar.acl.list$parameters$showDeleted$description
#> [1] "Whether to include deleted ACLs in the result. Deleted ACLs are represented by role equal to \"none\". Deleted ACLs will always be included if syncToken is provided. Optional. The default is False."
#> 
#> $calendar.acl.list$parameters$showDeleted$location
#> [1] "query"
#> 
#> 
#> $calendar.acl.list$parameters$syncToken
#> $calendar.acl.list$parameters$syncToken$type
#> [1] "string"
#> 
#> $calendar.acl.list$parameters$syncToken$description
#> [1] "Token obtained from the nextSyncToken field returned on the last page of results from the previous list request. It makes the result of this list request contain only entries that have changed since then. All entries deleted since the previous list request will always be in the result set and it is not allowed to set showDeleted to False.\nIf the syncToken expires, the server will respond with a 410 GONE response code and the client should clear its storage and perform a full synchronization without any syncToken.\nLearn more about incremental synchronization.\nOptional. The default is to return all entries."
#> 
#> $calendar.acl.list$parameters$syncToken$location
#> [1] "query"
#> 
#> 
#> $calendar.acl.list$parameters$alt
#> $calendar.acl.list$parameters$alt$type
#> [1] "string"
#> 
#> $calendar.acl.list$parameters$alt$description
#> [1] "Data format for the response."
#> 
#> $calendar.acl.list$parameters$alt$default
#> [1] "json"
#> 
#> $calendar.acl.list$parameters$alt$enum
#> [1] "json"
#> 
#> $calendar.acl.list$parameters$alt$enumDescriptions
#> [1] "Responses with Content-Type of application/json"
#> 
#> $calendar.acl.list$parameters$alt$location
#> [1] "query"
#> 
#> 
#> $calendar.acl.list$parameters$fields
#> $calendar.acl.list$parameters$fields$type
#> [1] "string"
#> 
#> $calendar.acl.list$parameters$fields$description
#> [1] "Selector specifying which fields to include in a partial response."
#> 
#> $calendar.acl.list$parameters$fields$location
#> [1] "query"
#> 
#> 
#> $calendar.acl.list$parameters$key
#> $calendar.acl.list$parameters$key$type
#> [1] "string"
#> 
#> $calendar.acl.list$parameters$key$description
#> [1] "API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token."
#> 
#> $calendar.acl.list$parameters$key$location
#> [1] "query"
#> 
#> 
#> $calendar.acl.list$parameters$oauth_token
#> $calendar.acl.list$parameters$oauth_token$type
#> [1] "string"
#> 
#> $calendar.acl.list$parameters$oauth_token$description
#> [1] "OAuth 2.0 token for the current user."
#> 
#> $calendar.acl.list$parameters$oauth_token$location
#> [1] "query"
#> 
#> 
#> $calendar.acl.list$parameters$prettyPrint
#> $calendar.acl.list$parameters$prettyPrint$type
#> [1] "boolean"
#> 
#> $calendar.acl.list$parameters$prettyPrint$description
#> [1] "Returns response with indentations and line breaks."
#> 
#> $calendar.acl.list$parameters$prettyPrint$default
#> [1] "true"
#> 
#> $calendar.acl.list$parameters$prettyPrint$location
#> [1] "query"
#> 
#> 
#> $calendar.acl.list$parameters$quotaUser
#> $calendar.acl.list$parameters$quotaUser$type
#> [1] "string"
#> 
#> $calendar.acl.list$parameters$quotaUser$description
#> [1] "An opaque string that represents a user for quota purposes. Must not exceed 40 characters."
#> 
#> $calendar.acl.list$parameters$quotaUser$location
#> [1] "query"
#> 
#> 
#> $calendar.acl.list$parameters$userIp
#> $calendar.acl.list$parameters$userIp$type
#> [1] "string"
#> 
#> $calendar.acl.list$parameters$userIp$description
#> [1] "Deprecated. Please use quotaUser instead."
#> 
#> $calendar.acl.list$parameters$userIp$location
#> [1] "query"
#> 
#> 
#> 
#> $calendar.acl.list$scopes
#> [1] "calendar"
#> 
#> $calendar.acl.list$description
#> [1] "Returns the rules in the access control list for the calendar."
#> 
#> $calendar.acl.list$response
#> [1] "Acl"
#> 
#> $calendar.acl.list$parameterOrder
#> [1] "calendarId"
#> 
#> $calendar.acl.list$supportsSubscription
#> [1] TRUE
#> 
#>