Functions that are no longer used.

These functions were available in a very early version of the package. Please use these replacement functions instead:

calendar_auth() and calendar_token()

calendar_list()

calendar_events()

get_google_token(google_app)

get_gcal_list(token = calendar_token())

get_gcal_events(
  id,
  token = calendar_token(),
  days_in_past = 90,
  days_in_future = 90,
  now = Sys.Date(),
  max_results = 250
)

Arguments

google_app
No longer used

token
token obtained from calendar_token()

id
calendar id, obtained from calendar_list()

days_in_past
Restrict results to date range, number of days in past

days_in_future
Restrict results to date range, number of days into future

now
Reference time stamp, defaulting to Sys.time()

max_results
Used internally to specify the maximum number of results retrieved per "page" of the query.